LayoutTests/http/tests/quicklook/submit-form-blocked.html

<!DOCTYPE html>
<html>
<head>
<script src="resources/tap-at-point-and-notify-done.js"></script>
<script>
if (window.testRunner) {
    testRunner.dumpAsText();
    testRunner.waitUntilDone();
}
</script>
</head>
<body>
<p>This test verifies that form submission is blocked using QuickLook to preview a Microsoft Word document. This test PASSED only if a security error is logged to the console.</p>
<iframe src="resources/submit-form-blocked.docx" onload="tapAtPointAndNotifyDone(this.offsetLeft + 5, this.offsetTop + 5)"></iframe>
</body>
</html>
