LayoutTests/platform/ios/http/tests/quicklook/submit-form-blocked-expected.txt

CONSOLE MESSAGE: line 2: Blocked form submission to 'http://127.0.0.1:8000/quicklook/resources/fail.html' because the form's frame is sandboxed and the 'allow-forms' permission is not set.
This test verifies that form submission is blocked using QuickLook to preview a Microsoft Word document. This test PASSED only if a security error is logged to the console.
